provide an input for the post url and the post auth key
as an mvp, this will just post whatever is hit as content, and send it with
whatever was received from the last response. The post service needs to prefer
the local id, but the service uid to facilitate this.

Alternately, the client could only send the id instead of both.