provide a means to write the data to localStorage
for now, this is quality gate closed until an id mechanism can be determined.
Because id is currently not overwritten at the service, perhaps a cid or some
other attribute needs to be appended to metadata to allow mapping. In addition,
some version of history needs to be retained locally for replays as well.